- fix udelay prototype

+++ nash/lib.h
@@ -68,6 +68,6 @@ extern int getDevNumFromProc(char * file, char * device);
 
 extern int stringsort(const void *v0, const void *v1);
 
-extern void udelay(int usecs);
+extern void udelay(long long usecs);
 
 #endif /* NASH_LIB_H */
